requirements.txt
flask==2.3.3
pandas==2.2.3
python-dotenv==1.0.1
playwright==1.49.0
openpyxl==3.1.5
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4